combine non-simple methods ; add ThukralXB methods
