+++ src/derivative_free.jl
@@ -115,100 +115,6 @@ function update_state(method::Order1, F, o::AbstractUnivariateZeroState{T,S}, op
 
 end
 
-##################################################
-
-
-"""
-    Roots.Order1B()
-    Roots.King()
-
-A superlinear (order 1.6...) modification of the secant method for multiple roots.
-Presented in A SECANT METHOD FOR MULTIPLE ROOTS, by RICHARD F. KING, BIT 17 (1977), 321-328
-
-The basic idea is similar to Schroder's method: apply the secant method
-to  `f/f'`. However, this uses `f' ~ fp = (fx - f(x-fx))/fx` (a Steffensen step). In
-this implementation, `Order1B`, when `fx` is too big, a single secant step of `f`
-is used.
-
-The *asymptotic* error, `eᵢ = xᵢ - α`, is given by
-`eᵢ₊₂ = 1/2⋅G''/G'⋅ eᵢ⋅eᵢ₊₁ + (1/6⋅G'''/G' - (1/2⋅G''/G'))^2⋅eᵢ⋅eᵢ₊₁⋅(eᵢ+eᵢ₊₁)`.
-
-"""
-struct King <: AbstractSecant end
-struct Order1B <: AbstractSecant end
-
-struct KingState{T,S} <: AbstractUnivariateZeroState{T,S}
-    xn1::T
-    xn0::T
-    fxn1::S
-    fxn0::S
-    G0::S
-end
-
-
-function init_state(M::Union{King, Order1B}, F, x₀, x₁::T, fx₀, fx₁::S) where {T,S}
-    fₛ₀ = F(x₀ - fx₀*oneunit(T)/oneunit(S))
-    G₀ = -fx₀^2 / (fₛ₀ - fx₀)
-    KingState(x₁, x₀, fx₁, fx₀, G₀)
-end
-initial_fncalls(::Union{King, Order1B}) = 3
-
-function update_state(method::Order1B, F, o::KingState{T,S}, options, l=NullTracks())  where {T, S}
-    if do_guarded_step(Order1B(), o)
-        state, flag = update_state(Order1(), F, o, options, l)
-
-        x0,fx0 = state.xn0, state.fxn0 # clunky! Need to update G₀ after Order1() step
-        fₛ = F(x0 - fx0*oneunit(T)/oneunit(S))
-        incfn(l)
-        G₀ = -fx0^2 / (fₛ - fx0)
-        @set! state.G0 = G₀
-
-        return (state, flag)
-    else
-        update_state(King(), F, o, options, l)
-    end
-end
-
-
-function update_state(method::King, F,
-                      o::KingState{T,S}, options, l=NullTracks())  where {T, S}
-
-
-    x0, x1 = o.xn0, o.xn1
-    fx0, fx1 = o.fxn0, o.fxn1
-    G₀ = o.G0
-    fₛ₁ = F(x1 - fx1*oneunit(T)/oneunit(S))
-    incfn(l)
-    G₁ = -fx1^2 / (fₛ₁ - fx1)
-
-    m = (x1-x0) / (G₁-G₀) # approximate value of `m`, the multiplicity
-
-    if abs(m) <= 1e-2 * oneunit(m)
-        log_message(l, "Estimate for multiplicity has issues. ")
-        return (o, true)
-    end
-
-    Δ = G₁ * (x1 - x0) / (G₁ - G₀)
-
-    if isissue(Δ)
-        log_message(l, "Increment `Δx` has issues. ")
-        return o, true
-    end
-
-    x0, x1 = x1, x1 - Δ
-    fx0, fx1 = fx1, F(x1)
-    incfn(l)
-
-    @set! o.xn0 = x0
-    @set! o.xn1 = x1
-    @set! o.fxn0 = fx0
-    @set! o.fxn1 = fx1
-    @set! o.G0 = G₁
-
-
-    return o, false
-
-end
 
 
 
@@ -271,92 +177,6 @@ function update_state(method::Steffensen, F, o::AbstractUnivariateZeroState{T,S}
     @set! o.fxn1 = fx1
 
     return o, false
-end
-
-##################################################
-
-### Order2B() Esser method
-"""
-    Roots.Order2B()
-    Roots.Esser()
-
-Esser's method. This is a quadratically convergent method that, like
-Schroder's method, does not depend on the multiplicity of the
-zero. Schroder's method has update step `x - r2/(r2-r1) * r1`, where `ri =
-fⁱ⁻¹/fⁱ`. Esser approximates `f' ~ f[x-h, x+h], f'' ~
-f[x-h,x,x+h]`, where `h = fx`, as with Steffensen's method, Requiring 3
-function calls per step. The implementation `Order2B` uses a secant
-step when `|fx|` is considered too large.
-
-
-Esser, H. Computing (1975) 14: 367. DOI: [10.1007/BF02253547](https://doi.org/10.1007/BF02253547)
-Eine stets quadratisch konvergente Modifikation des Steffensen-Verfahrens
-
-
-Example
-```
-f(x) = cos(x) - x
-g(x) = f(x)^2
-x0 = pi/4
-find_zero(f, x0, Order2(), verbose=true)        #  3 steps / 7 function calls
-find_zero(f, x0, Roots.Order2B(), verbose=true) #  4 / 9
-find_zero(g, x0, Order2(), verbose=true)        #  22 / 45
-find_zero(g, x0, Roots.Order2B(), verbose=true) #  4 / 10
-```
-"""
-struct Esser <: AbstractSecant end
-struct Order2B <: AbstractSecant end
-
-function update_state(method::Order2B, fs, o::AbstractUnivariateZeroState{T,S}, options, l=NullTracks())  where {T, S}
-     update_state_guarded(method, Secant(), Esser(), fs, o, options, l)
-end
-
-function update_state(method::Esser, F,
-                      o::AbstractUnivariateZeroState{T,S}, options, l=NullTracks())  where {T, S}
-
-
-    x1, fx1 = o.xn1, o.fxn1
-
-    f0 = fx1
-
-    f1  = F(x1 + f0 * oneunit(T) / oneunit(S))
-    f_1 = F(x1 - f0 * oneunit(T) / oneunit(S))
-    incfn(l, 2)
-
-    # h = f0
-    # r1 = f/f' ~ f/f[x+h,x-h]
-    # r2 = f'/f'' ~ f[x+h, x-h]/f[x-h,x,x+h]
-    r1 = f0 * 2*f0 / (f1 - f_1) * oneunit(T) / oneunit(S)
-    r2 = (f1 - f_1)/(f1 - 2*f0 + f_1) * f0/2 * oneunit(T) / oneunit(S)
-
-    k = r2/(r2-r1)  # ~ m
-
-    if abs(k) <= 1e-2 * oneunit(k)
-        log_message(l, "Estimate for multiplicity had issues. ")
-        return o, true
-#        #@info "estimate for m is *too* small"
-    end
-
-    delta = k * r1
-
-    if isissue(delta)
-        log_message(l, "Increment `Δx` has issues. ")
-        return o, true
-    end
-
-    x0, x1 = x1, x1 - delta
-    fx0, fx1 = fx1, F(x1)
-    incfn(l)
-
-
-
-    @set! o.xn0 = x0
-    @set! o.xn1 = x1
-    @set! o.fxn0 = fx0
-    @set! o.fxn1 = fx1
-
-    return o, false
-
 end
